network: extract tls module from connection
As the tls module got quite large, it makes sense to extract it to the
network supermodule.
