+++ scylla/src/client/session.rs
@@ -19,7 +19,7 @@ use crate::errors::{
 };
 use crate::frame::response::result;
 #[cfg(feature = "__tls")]
-use crate::network::TlsProvider;
+use crate::network::tls::TlsProvider;
 use crate::network::{Connection, ConnectionConfig, PoolConfig, VerifiedKeyspaceName};
 use crate::observability::driver_tracing::RequestSpan;
 use crate::observability::history::{self, HistoryListener};
